+++ src/apps/lwaftr/lwcounter.lua
@@ -40,15 +40,18 @@ end
 
 -- Temporary hack until we migrate all of the lwAFTR's apps to use the
 -- SHM frame mechanism.
-local ipv4_reassemble = require('apps.ipv4.reassemble')
-local ipv6_reassemble = require('apps.ipv6.reassemble')
+local managed_counters = {}
+for _, app in ipairs({ require('apps.ipv4.reassemble').Reassembler,
+                       require('apps.ipv6.reassemble').Reassembler,
+                       require('apps.ipv4.fragment').Fragmenter }) do
+   for k, v in pairs(app.shm) do managed_counters[k] = true end
+end
+
 function init_counters ()
    local counters = {}
    for k, id in pairs(counter_names()) do
-      if ipv4_reassemble.Reassembler.shm[k] then
-         -- Don't create this counter; IPv4 reassemble app handles it.
-      elseif ipv6_reassemble.Reassembler.shm[k] then
-         -- Don't create this counter; IPv6 reassemble app handles it.
+      if managed_counters[k] then
+         -- Don't create this counter; it's managed by an app.
       else
          counters[k] = {counter}
       end
